Consolidate PlayReady into a single session class
